Switched “i” field of IndexVal and IQIndexVal to be of type long. Switched int to long in other places in index and iqindex code.

+++ itensor/index.cc
@@ -200,7 +200,7 @@ IndexVal()
     { }
 
 IndexVal::
-IndexVal(const Index& index_, int i_) 
+IndexVal(const Index& index_, long i_) 
     : 
     index(index_),
     i(i_)
